+++ contrib/rgb565_to_png.html
@@ -1,0 +1,113 @@
+<!DOCTYPE html>
+<html lang="de">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>RGB565 Hex zu PNG</title>
+    <script src="https://cdn.jsdelivr.net/npm/heatshrink-js@1.0.0/heatshrink.min.js"></script>
+</head>
+<body>
+    <h1>RGB565 Hex zu PNG</h1>
+    
+    <label for="width">Breite:</label>
+    <input type="number" id="width" name="width" min="1" value="100">
+    
+    <label for="height">Höhe:</label>
+    <input type="number" id="height" name="height" min="1" value="100">
+    
+    <br>
+    
+    <label for="hexFormat">Hex-Format:</label>
+    <select id="hexFormat" name="hexFormat">
+        <option value="0x">0x</option>
+        <option value="">Ohne Präfix</option>
+    </select>
+    
+    <label for="separator">Trennzeichen:</label>
+    <select id="separator" name="separator">
+        <option value=" ">Leerzeichen</option>
+        <option value=",">Komma</option>
+        <option value=";">Semikolon</option>
+        <option value="">Keines</option>
+    </select>
+    
+    <br>
+    
+    <label for="useHeatshrink">Heatshrink Dekomprimierung:</label>
+    <input type="checkbox" id="useHeatshrink" name="useHeatshrink">
+    
+    <br>
+    
+    <textarea id="hexInput" rows="10" cols="50" placeholder="Geben Sie die Hex-Werte im RGB565-Format ein..."></textarea>
+    <br>
+    <button onclick="convertHexToCanvas()">Konvertieren</button>
+    
+    <h2>Ausgabe</h2>
+    <canvas id="canvas"></canvas>
+    
+    <script>
+        function hexToBytes(hex) {
+            let bytes = [];
+            for (let c = 0; c < hex.length; c += 2) {
+                bytes.push(parseInt(hex.substr(c, 2), 16));
+            }
+            return bytes;
+        }
+
+        function bytesToHex(bytes) {
+            return bytes.map(byte => byte.toString(16).padStart(2, '0')).join('');
+        }
+
+        function decompressHeatshrink(hexInput) {
+            const compressedBytes = hexToBytes(hexInput);
+            const decompressedBytes = heatshrink.decompress(compressedBytes);
+            return bytesToHex(decompressedBytes);
+        }
+
+        function convertHexToCanvas() {
+            const width = parseInt(document.getElementById('width').value);
+            const height = parseInt(document.getElementById('height').value);
+            const hexFormat = document.getElementById('hexFormat').value;
+            const separator = document.getElementById('separator').value;
+            const useHeatshrink = document.getElementById('useHeatshrink').checked;
+            let hexInput = document.getElementById('hexInput').value;
+            
+            if (useHeatshrink) {
+                hexInput = decompressHeatshrink(hexInput);
+            }
+
+            // Split the input based on the selected separator
+            const hexValues = hexInput.split(separator).map(hex => hex.replace(hexFormat, '').trim()).filter(Boolean);
+            
+            if (hexValues.length !== width * height) {
+                alert('Die Anzahl der Hex-Werte entspricht nicht der angegebenen Bildgröße.');
+                return;
+            }
+            
+            const canvas = document.getElementById('canvas');
+            canvas.width = width;
+            canvas.height = height;
+            const ctx = canvas.getContext('2d');
+            const imageData = ctx.createImageData(width, height);
+            const data = imageData.data;
+            
+            for (let i = 0; i < hexValues.length; i++) {
+                const hex = hexValues[i];
+                const rgb565 = parseInt(hex, 16);
+                
+                const r = ((rgb565 >> 11) & 0x1F) << 3;
+                const g = ((rgb565 >> 5) & 0x3F) << 2;
+                const b = (rgb565 & 0x1F) << 3;
+                
+                const index = i * 4;
+                data[index] = r;
+                data[index + 1] = g;
+                data[index + 2] = b;
+                data[index + 3] = 255; // Vollständige Deckkraft
+            }
+            
+            ctx.putImageData(imageData, 0, 0);
+        }
+    </script>
+</body>
+</html>
