--- src/credentials.h.template
@@ -1,7 +1,0 @@
-#ifndef WIFI_CREDENTIALS_H
-#define WIFI_CREDENTIALS_H
-
-const char* ssid = "SSID_HIER_EINFUEGEN";
-const char* password = "PASSWORT_HIER_EINFUEGEN";
-const char *mqtt_server = "192.168.1.99";
-#endif
